requirements.txt
appdirs==1.4.4
astroid==2.4.2
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
colorama==0.4.4
cryptography==3.4.7
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.7.0
keyboard==0.13.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.1
password-validator==1.0
pickleshare==0.7.5
pipenv==2020.11.15
prompt-toolkit==3.0.8
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
Pygments==2.7.3
pyotp==2.6.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pywin32==300
pyzmq==20.0.0
six==1.15.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
virtualenv==20.2.2
virtualenv-clone==0.5.4
wcwidth==0.2.5
wrapt==1.12.1
